% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{AsBackwardTime}
\alias{AsBackwardTime}
\title{Convert Ordered Time as Time to Last Time Point}
\usage{
AsBackwardTime(time)
}
\arguments{
\item{time}{A numeric vector of time.}
}
\value{
A list consisting of a numeric vector of time with respect to last
time point and last time point.
}
\description{
Convert ordered time as time to last timepoint.
}
