% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{AsForwardTime}
\alias{AsForwardTime}
\title{Convert Time to Last Timepoint as Normal Time}
\usage{
AsForwardTime(time, last.time)
}
\arguments{
\item{time}{A numeric vector of time.}

\item{last.time}{Actual time of last timepoint}
}
\value{
A vector of time in normal order
}
\description{
Convert time to last timepoint as normal time.
}
