% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{BNPR_to_df}
\alias{BNPR_to_df}
\title{Convert BNPR object summary statistics to data frame}
\usage{
BNPR_to_df(BNPR_obj, label, last_time)
}
\arguments{
\item{BNPR_obj}{A BNPR object}

\item{label}{Label to be displayed on the plot}

\item{last_time}{Last sampling time}
}
\description{
Convert BNPR object summary statistics to data frame
}
