% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{compare.Forecast}
\alias{compare.Forecast}
\title{Compare performance of two models by AE and W}
\usage{
\method{compare}{Forecast}(x, y, xlabel, ylabel, truth,
  color = c("#0083C3", "#EB975F"))
}
\arguments{
\item{x, }{y Two Forecast object.}
}
\value{
A ggplot object comparing the AE and W between the two methods.
}
\description{
Compare performance of two models by AE and W
}
