% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{ComparePerformance}
\alias{ComparePerformance}
\title{Compare performance of two models by AE and W}
\usage{
ComparePerformance(x, y, truth, method = c(1, 2))
}
\arguments{
\item{x, }{y A data frame returned by forecast function}
}
\value{
A ggplot object comparing the AE and W between the two methods
}
\description{
Compare performance of two models by AE and W
}
