% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{compute_peak_week}
\alias{compute_peak_week}
\title{Compute peakk week and peak ILI for each season}
\usage{
compute_peak_week(task_list, models, task_season_range, n = 100)
}
\arguments{
\item{task_list}{A list of tasks}

\item{models}{A list of INLA model for each task}

\item{n}{Number of posterior predictive samples}
}
\value{
A list containing corresponding peak week and peak ILI for each mode
}
\description{
Compute peakk week and peak ILI for each season
}
