% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{compute_truncation_time}
\alias{compute_truncation_time}
\title{Compute the date of the Monday of the week for truncation}
\usage{
compute_truncation_time(year, week)
}
\arguments{
\item{year}{The year of the truncation time}

\item{week}{The week of the truncation time}
}
\description{
Compute the date of the Monday of the week for truncation
}
