% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{computeAEW}
\alias{computeAEW}
\title{Compute absolute error and width between forecast and truth.}
\usage{
computeAEW(x, truth)
}
\description{
Compute absolute error and width between forecast and truth.
}
