% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ComputePropCnst}
\alias{ComputePropCnst}
\title{Compute Proportionality Constant}
\usage{
ComputePropCnst(n, traj, lim)
}
\arguments{
\item{n}{Number of samples.}

\item{traj}{A function of time (intensity).}

\item{lim}{A numeric tuple of start and end time.}
}
\value{
A numeric scalar as proportionality constant.
}
\description{
Compute proportionality constant to produce a correct expected number of
samples
}
