% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{create_data_avail_table}
\alias{create_data_avail_table}
\title{Create data availability table}
\usage{
create_data_avail_table(data)
}
\arguments{
\item{data}{A PILAF object}
}
\value{
A data frame of available data for each season
}
\description{
Create data availability table
}
