% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{create_grid}
\alias{create_grid}
\title{Create grid of time and week}
\usage{
create_grid(samp_times, coal_times, last_time, pred)
}
\arguments{
\item{samp_times}{Sampling times; the last sampling time is set to 0}

\item{coal_times}{Coalescent times; the same reference as samp_times}

\item{last_time}{The last time point}

\item{pred}{Integer; the number of week to forecast ahead}
}
\value{
A list containing the training and testing time and week
}
\description{
Create grid of time and week
}
