% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{create_season_task_list}
\alias{create_season_task_list}
\title{Create a list of tasks to run for each season}
\usage{
create_season_task_list(season_list, starting_time, start_week, end_week,
  min_pred)
}
\arguments{
\item{season_list}{A data frame of the start and end time of train and predict}

\item{start_week}{Starting week number}

\item{end_week}{Ending week number}

\item{min_pred}{Minimum number of weeks to predict}

\item{start_time}{Starting time of the data}
}
\value{
A data frame of tasks to run
}
\description{
Create a list of tasks to run for each season
}
