% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{find_peak_week}
\alias{find_peak_week}
\title{Find the peak week amd peak ILI from samples}
\usage{
find_peak_week(posterior_sample, task_entry, season_range)
}
\arguments{
\item{posterior_sample}{Asample from  inla.posterior.sample}

\item{task_entry}{A task entry corresponding to the sample}

\item{season_range}{The start and end of the season}
}
\value{
A data frame containing the peak week and peak ILI from the sample
}
\description{
Find the peak week amd peak ILI from samples
}
