% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{find_season_range}
\alias{find_season_range}
\title{Find the starting and ending time of a season}
\usage{
find_season_range(task_list)
}
\arguments{
\item{task_list}{A data frame of task generated from generate_task_list}
}
\value{
A data frame containing the start and end time of each season
}
\description{
Find the starting and ending time of a season
}
