% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{forecast.PILAF.count}
\alias{forecast.PILAF.count}
\title{Forecast using Only Count}
\usage{
\method{forecast}{PILAF.count}(x, formula)
}
\arguments{
\item{x}{A PILAF object with NA for values to forecast.}
}
\value{
A PILAF object with forecasted ILI counts and 95 % BCI
}
\description{
Forecast using Only Count
}
