% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{forecast.PILAF.joint}
\alias{forecast.PILAF.joint}
\title{Forecast using Count and Coalescent events}
\usage{
\method{forecast}{PILAF.joint}(x, formula = NULL, verbose = F)
}
\arguments{
\item{x}{A PILAF object with NA for values to forecast.}

\item{formula}{A INLA formula.}
}
\value{
A PILAF object with forecasted ILI counts and 95 % BCI.
}
\description{
Forecast using Count and Coalescent events
}
