% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{Forecast}
\alias{Forecast}
\title{Constructor to Forecast object}
\usage{
Forecast(time = integer(), week = integer(), mean = double(),
  quant0.025 = double(), quant0.975 = double(), iter = integer())
}
\arguments{
\item{time}{A numeric vector of time.}

\item{mean}{A numeric vector of forecast mean.}

\item{quant0.025}{A numeric vector of 0.025 quantile.}

\item{quant0.975}{A numeric vector 0.975 quantile.}

\item{iter}{A numeric vector to indicate group of trajectories.}
}
\value{
A Forecast object.
}
\description{
Constructor method to create an object from Forecast class
}
