% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{forecast_starting}
\alias{forecast_starting}
\title{Forecast starting at given time}
\usage{
forecast_starting(tree, last_time, week_start, year_start, formula,
  pred = 4, pref = TRUE)
}
\arguments{
\item{tree}{A phylo object}

\item{last_time}{The last sampling time in the tree}

\item{week_start}{The week of last timepoint for training}

\item{year_start}{The year of last timepoint for training}

\item{formula}{The formula for BNPR forecast}

\item{pred}{The number of weeks to forecast}

\item{pref}{Whether using preferential sampling}
}
\value{
A INLA object for forecast
}
\description{
Forecast starting at given time
}
