% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{generate_task_list}
\alias{generate_task_list}
\title{Generate task list}
\usage{
generate_task_list(data, start_week = 40, end_week = 22,
  min_pred = 4)
}
\arguments{
\item{data}{A data frame of data available}

\item{start_week}{Starting week for prediction of each season}

\item{end_week}{Ending week for prediction of each season}

\item{min_pred}{Minimum number of week to predict}
}
\value{
A data frame of tasks to run
}
\description{
Generate task list
}
