% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Get summary statistics from samples}
\usage{
get_summary_stats(samples)
}
\arguments{
\item{samples}{A numeric vector of samples to summarize}
}
\value{
A data frame containing the 95% BCI, mean and median of the samples
}
\description{
Get summary statistics from samples
}
