% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{is.Forecast}
\alias{is.Forecast}
\title{Check if an Object is Forecast}
\usage{
is.Forecast(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar indicating whether the object belongs to Forecast class.
}
\description{
Check if an object is of Forecast class
}
