% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{is.PILAF}
\alias{is.PILAF}
\title{Check if an Object is PILAF}
\usage{
is.PILAF(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar indicating whether the object belongs to PILAF class.
}
\description{
Check if an object is of PILAF class
}
