% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{LOSO_forecast}
\alias{LOSO_forecast}
\title{Leave one season out evaluation}
\usage{
LOSO_forecast(data, task_list, method = "count", formula = NULL)
}
\arguments{
\item{data}{A PILAF object}

\item{task_list}{A data frame of tasks to run}

\item{method}{Method used for forecasting (count or joint)}

\item{formula}{Model formula}
}
\value{
A list of prediction mean and 95% BCI
}
\description{
Leave one season out evaluation
}
