% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mean_coverage}
\alias{mean_coverage}
\title{Mean coverage}
\usage{
mean_coverage(lwr, upr, truth)
}
\arguments{
\item{lwr}{95 % BCI lower bound of forecast}

\item{upr}{95 % BCI upper bound of forecast}

\item{truth}{Real values}
}
\description{
Mean coverage
}
