% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{PILAF}
\alias{PILAF}
\title{Constructor to PILAF object}
\usage{
PILAF(time, week, coal, samp, ILI, coal.E, samp.E, ILI.E, iter)
}
\arguments{
\item{time}{A numeric vector of time.}

\item{week}{A numeric vector of week.}

\item{coal}{A numeric vector of coalescent event counts.}

\item{samp}{A numeric vector of sampling event counts.}

\item{ILI}{A numeric vector ILI counts.}

\item{coal.E}{A numeric vector of offset to coal.}

\item{samp.E}{A numeric vector of offset to samp.}

\item{ILI.E}{A numeric vector of offset to ILI.}

\item{iter}{A numeric vector to indicate group of trajectories.}
}
\value{
A PILAF object.
}
\description{
Constructor method to create an object from PILAF class
}
