% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{plot.forecast}
\alias{plot.forecast}
\title{Plot the Forecast Results}
\usage{
\method{plot}{forecast}(x, pilaf, truth = NULL)
}
\arguments{
\item{x}{A PILAF object.}

\item{truth}{An optional PILAF object that shows the truth.}

\item{forecast}{A forecast object.}
}
\value{
A ggplot object for the forecasting results.
}
\description{
Plot the forecasting results
}
