% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_MCC_Ne}
\alias{plot_MCC_Ne}
\title{Plot MCC tree and skygrid effective population size}
\usage{
plot_MCC_Ne(MCC, root_time, last_time, Ne, breaks, main = NULL,
  ratio = c(1.2, 1), mask_range = NULL, ylimits = NULL)
}
\arguments{
\item{MCC}{A nexus object}

\item{root_time}{The actual time for the root of the tree}

\item{last_time}{The last time to be plotted}

\item{Ne}{A list of data frame containing Ne summary statistics}
}
\description{
Plot MCC tree and skygrid effective population size
}
