% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_traj_heatmap}
\alias{plot_traj_heatmap}
\title{plot real CDC time series data}
\usage{
plot_traj_heatmap(data, value)
}
\arguments{
\item{data}{A data frame containing year, week and total ILI counts}

\item{value}{The name of the variable that contains ILI counts}
}
\description{
plot real CDC time series data
}
