% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot a Ne trajectory}
\usage{
plot_trajectory(df, ylimits, breaks, root_time, vline = NULL)
}
\arguments{
\item{df}{A data frame containing six columns (Time, Mean, Median, Upper, Lower, Label)}

\item{ylimits}{range of y axis}

\item{breaks}{Breaks for x axis}

\item{root_time}{Minimum for x axis}

\item{vline}{x intercept for a dashed line to denote start of forecast}
}
\value{
A ggplot object of the trajectory
}
\description{
Plot a Ne trajectory
}
