% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{PlotSimulations}
\alias{PlotSimulations}
\title{Plot Simulations}
\usage{
PlotSimulations(sim)
}
\arguments{
\item{sim}{A data frame from PILAF::SimulateILISampCoalCountsN}
}
\description{
Plot simulated counts.
}
