% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{SimulateCoalCounts}
\alias{SimulateCoalCounts}
\title{Simulate Coalescent Counts}
\usage{
SimulateCoalCounts(grid, samp.times, samp.num, traj)
}
\arguments{
\item{grid}{A numeric vector of grid.}

\item{samp.times}{A numeric vector of sampling times.}

\item{samp.num}{A numeric vector of number of samples per time.}

\item{traj}{A function of time.}
}
\value{
A data frame of coalescent counts.
}
\description{
Simulate coalescent counts given sampling events and intensity.
}
\seealso{
phylodyn::coalsim
}
