% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{SimulateILISampCoalCounts}
\alias{SimulateILISampCoalCounts}
\title{Simulate ILI, Sampling, and Coalescent counts}
\usage{
SimulateILISampCoalCounts(lim, flu.Ne, flu.sampNum, ILI.sampNum)
}
\arguments{
\item{lim}{A numeric tuple of start and end times.}

\item{flu.Ne}{A function of time (for flu effective popluation size).}

\item{flu.sampNum}{A numeric scalar of expected number of flu samples in lim.}

\item{ILI.sampNum}{A numeric scalar of expected number of sampling events.
in lim}
}
\value{
A list containing coalescent, sampling and ILI counts
}
\description{
Simulate flu sampling, coalescence and ILI counts given effective population
size trajectory and number of expected samples.
}
