% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{SMAPE}
\alias{SMAPE}
\title{Symmetric mean absolute percentage error}
\usage{
SMAPE(x, truth)
}
\arguments{
\item{x}{Forecast values}

\item{truth}{Real values}
}
\description{
Symmetric mean absolute percentage error
}
