% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILAF.R
\name{summary.forecast}
\alias{summary.forecast}
\title{Summarize the Forecast}
\usage{
\method{summary}{forecast}(x, truth)
}
\arguments{
\item{x}{A forecast object.}

\item{truth}{A data frame containing time, true ILI.}
}
\value{
A data frame containing the absolute error of forecast.
}
\description{
Summarize the forecast by pointwise mean absolute error and mean relative
width.
}
