% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{truncate_BNPR_PS}
\alias{truncate_BNPR_PS}
\title{Fit BNPR_PS with truncated data}
\usage{
truncate_BNPR_PS(year_start, week_start, tree, last_time, n = 1,
  include_trunc = T, thresh = 0.019, ...)
}
\arguments{
\item{year_start}{The year of the truncation time}

\item{week_start}{The week of the truncation time}

\item{tree}{The phylo object}

\item{last_time}{The last sampling time}

\item{n}{The number of samples near the truncation time}

\item{include_trunc}{Whether to include the truncation timepoint}

\item{thresh}{Maximum distance from truncation time of the n samples}

\item{...}{arguments to BNPR_PS}
}
\description{
Fit BNPR_PS with truncated data
}
