% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{truncate_data}
\alias{truncate_data}
\title{Truncates a phylogentic tree}
\usage{
truncate_data(tree, truncation_time)
}
\arguments{
\item{tree}{a phylo object}

\item{truncation_time}{The time of last sampling point}
}
\description{
Truncates a phylogentic tree
}
