% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualizations.R
\name{visualize_flu_season}
\alias{visualize_flu_season}
\title{Visualize the time series for each Flu Season}
\usage{
visualize_flu_season(data, time_series_name, subset_seasons = NULL)
}
\arguments{
\item{data}{A data frame contain three columns (year, week, name_of_time_series)}

\item{time_series_name}{The column name for the time series}

\item{subset_seasons}{Seasons to keep}
}
\value{
A ggplot object visualizing from week 40 to week 21
}
\description{
Plot the time series for each flu season (week 40 - week 21).
}
