% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{YearlyLogisticTraj}
\alias{YearlyLogisticTraj}
\title{A Logistic Trajectory with Period 52}
\usage{
YearlyLogisticTraj(...)
}
\description{
A logistic trajectory (based on `phylodyn::logistic_traj`) with period 52.
}
