% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPR_forecast.R
\name{BNPR_forecast}
\alias{BNPR_forecast}
\title{BNPR forecast}
\usage{
BNPR_forecast(data, last_time, formula, lengthout = 100, pref = FALSE,
  prec_alpha = 0.01, prec_beta = 0.01, beta1_prec = 0.001,
  fns = NULL, log_fns = TRUE, simplify = TRUE, derivative = FALSE,
  forward = TRUE, pred = 4)
}
\description{
BNPR forecast
}
