% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{plot.Forecast}
\alias{plot.Forecast}
\title{Plot the Forecast Results}
\usage{
\method{plot}{Forecast}(x, pilaf, truth = NULL)
}
\arguments{
\item{x}{A PILAF object.}

\item{truth}{An optional PILAF object that shows the truth.}

\item{forecast}{A Forecast object.}
}
\value{
A ggplot object for the forecasting results.
}
\description{
Plot the forecasting results
}
